Add flashcard and flashcard pack tables to the database schema
